refactor: use sqrt from gnark-crypto
